src/app/requirements.txt
bcrypt
Flask-RESTful==0.3.7
Flask==2.2.2
gunicorn==20.0.4
mysql-connector-python
python-dotenv==0.19.0
flask-apispec==0.11.0
jinja2
themoviedb
justwatch
canistreamit
Werkzeug==2.2.2
wtforms
email_validator
